// front/src/interfaces/TypesRegister.ts
export interface valuesTypesRegister {
  name: string;
  email: string;
  password: string;
  confirm_password: string;
  phone: string;
  dni: string;
}

export interface valuesTypesRegisterPrueba {
  name: string;
  email: string;
  password: string;
  confirm_password: string;
  phone: string;
  dni: string;
}
export interface valuesTypesRegisterGoogle {
  name: string;
  email: string;
  password: string;
  confirm_password: string;
  phone: string;
  dni: string;
  authProvider: string;
}